Establish Home & NewTask Page, Make AppJS slimmer
