+++ src/pages/NewTask.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function NewTask() {
+  return <div>NEW TASK</div>;
+}
+
+export default NewTask;
